requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.4
    # via black
asgiref==3.3.1
    # via django
backcall==0.2.0
    # via ipython
black==20.8b1
    # via -r requirements.in
click==7.1.2
    # via black
decorator==4.4.2
    # via ipython
django-countries==7.1
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django==3.1.7
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
factory-boy==3.2.0
    # via -r requirements.in
faker==8.0.0
    # via factory-boy
ipython-genutils==0.2.0
    # via traitlets
ipython==7.22.0
    # via -r requirements.in
jedi==0.18.0
    # via ipython
mypy-extensions==0.4.3
    # via black
parso==0.8.2
    # via jedi
pathspec==0.8.1
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.18
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pygments==2.8.1
    # via ipython
python-dateutil==2.8.1
    # via faker
pytz==2021.1
    # via django
pyyaml==5.4.1
    # via -r requirements.in
regex==2021.3.17
    # via black
six==1.15.0
    # via python-dateutil
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via black
traitlets==5.0.5
    # via ipython
typed-ast==1.4.2
    # via black
typing-extensions==3.7.4.3
    # via black
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
